Cache values of the shbufsz routine since it's in the critical path and doesn't change; add MAXMSG_BFQ control
git-svn-id: svn://svn.code.sf.net/p/xymon/code@7779 44351d6e-118b-4698-b696-ce33095ecaa4

+++ branches/4.x-master/lib/xymond_buffer.h
@@ -13,6 +13,6 @@
 
 enum msgchannels_t { C_STATUS=1, C_STACHG, C_PAGE, C_DATA, C_NOTES, C_ENADIS, C_CLIENT, C_CLICHG, C_USER, C_FEEDBACK_QUEUE, C_LAST };
 
-extern unsigned int shbufsz(enum msgchannels_t chnid);
+extern size_t shbufsz(enum msgchannels_t chnid);
 #endif
